Update readme, rename helper function files
