+++ Readme.md
@@ -8,6 +8,4 @@ The goal of this bot is to create weirder and more challenging
 commands to practice problem solving with limited tools.
 
 ## <u>**Todo**</u>
-  - [ ] Move registering and deregistering commands to separate programs
-  - [x] Figure out command perms (permissions don't apply to commands for some reason)
   - [ ] Figure out reactions (reactions are not collected from messages)
